+++ tools/pseudoalign.cpp
@@ -52,14 +52,14 @@ int do_map(index_type const& index, fastx_parser::FastxParser<fastx_parser::Read
     if ((algo == pseudoalignment_algorithm::SKIPPING or
          algo == pseudoalignment_algorithm::SKIPPING_KALLISTO) and
         (index.get_dict())->canonicalized()) {
-        std::vector<uint32_t> unitig_ids;                           // for use with skipping
+        std::vector<uint64_t> unitig_ids;                           // for use with skipping
         std::vector<std::pair<projected_hits, int>> kallisto_hits;  // for use with kallisto psa
 
         piscem_psa::hit_searcher hs(&index);
         sshash::streaming_query_canonical_parsing qc(index.get_dict());
 
         auto get_hits_piscem_psa = [&qc, &hs](const std::string& seq,
-                                              std::vector<uint32_t>& unitig_ids) -> void {
+                                              std::vector<uint64_t>& unitig_ids) -> void {
             hs.clear();
             auto had_hits = hs.get_raw_hits_sketch(seq, qc, true, false);
             if (had_hits) {
@@ -71,7 +71,7 @@ int do_map(index_type const& index, fastx_parser::FastxParser<fastx_parser::Read
 
         auto get_hits_kallisto_psa = [&index, &kallisto_hits](
                                          const std::string& seq,
-                                         std::vector<uint32_t>& unitig_ids) -> void {
+                                         std::vector<uint64_t>& unitig_ids) -> void {
             kallisto_hits.clear();
             match(seq, seq.length(), &index, kallisto_hits);
             if (!kallisto_hits.empty()) {
